pyproject.toml
[project]
name = "mitre-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
]
